+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add minimzed testcases.
 - Add renaming of packages.
 - Add interface renaming.
+- Add renaming of `endmodule` labels.
 
 ## Changed
 - Use [sv-parser](https://github.com/dalance/sv-parser) as the main SV parser.